Added files to run pvpython to generate movie

+++ Dockerfile
@@ -1,9 +1,12 @@
 FROM ubuntu:18.04 AS mergepolydata
 
 RUN apt-get update && \
-  apt-get install -y  --no-install-recommends \
-  cmake git g++ ca-certificates vim make libgl1-mesa-dev libxt-dev xvfb && \
-  rm -rf /var/lib/apt/lists/*
+  apt-get install -y  --no-install-recommends curl python3-dev libpng-dev\
+  cmake git g++ ca-certificates vim make libgl1-mesa-dev libxt-dev \
+  zlib1g-dev && rm -rf /var/lib/apt/lists/*
+
+# Install CMake, direct binary
+RUN curl -s "https://cmake.org/files/v3.17/cmake-3.17.0-Linux-x86_64.tar.gz" | tar --strip-components=1 -xz -C /usr/local
 
 # Setup home environment
 RUN useradd -rm -d /home/ubuntu -s /bin/bash -g root -G sudo -u 1000 ubuntu
@@ -12,12 +15,17 @@ WORKDIR /home/ubuntu
 
 # Setup VTK
 RUN git clone -b 'v7.1.1' --single-branch https://github.com/Kitware/VTK.git
-RUN mkdir VTK/build;cd VTK/build;cmake ../ -DBUILD_SHARED_LIBS=OFF -DCMAKE_BUILD_TYPE=Release;make -j 8; cd ../..
+RUN mkdir VTK/build;cd VTK/build;cmake .. -DBUILD_SHARED_LIBS=OFF -DCMAKE_BUILD_TYPE=Release;make -j 16
 
 # Setup MergePolyData
 ADD https://api.github.com/repos/PSUCompBio/MergePolyData/git/refs/heads/develop version.json
 RUN git clone  -b develop --single-branch https://github.com/PSUCompBio/MergePolyData.git
-RUN mkdir MergePolyData/build;cd MergePolyData/build;cmake .. -DVTK_DIR=/home/ubuntu/VTK/build;make -j 8
+RUN mkdir MergePolyData/build;cd MergePolyData/build;cmake .. -DVTK_DIR=/home/ubuntu/VTK/build;make -j 16
+
+# Setup pvpython
+RUN git clone --recursive https://gitlab.kitware.com/paraview/paraview-superbuild.git
+RUN cd paraview-superbuild;git fetch origin;git checkout v5.8.0;git submodule update 
+RUN mkdir paraviewSuperbuildBuild;cd paraviewSuperbuildBuild;cmake -DENABLE_python=ON -DENABLE_ffmpeg=ON -DBUILD_SHARED_LIBS_paraview=OFF -DBUILD_TESTING=OFF -DENABLE_nlohmannjson=OFF -DENABLE_python3=ON -DENABLE_png=ON -DUSE_SYSTEM_python3=ON -DUSE_SYSTEM_zlib=ON -DUSE_SYSTEM_png=ON -DCMAKE_BUILD_TYPE=Release ../paraview-superbuild ;make -j 16 
 
 FROM ubuntu:18.04 AS multiviewport
 
@@ -40,6 +48,17 @@ COPY --from=mergepolydata ["/home/ubuntu/MergePolyData/build/examples/multipleVi
   "/home/ubuntu/MergePolyData/build/MultipleViewPorts", \
   "/home/ubuntu/MultiViewPortRun/"]
 
+RUN mkdir Paraview
+
+COPY --from=mergepolydata ["/home/ubuntu/paraviewSuperbuildBuild/install/bin/pvpython", \
+  "/home/ubuntu/paraviewSuperbuildBuild/install/lib/libavformat.so.55.48.100", \
+  "/home/ubuntu/paraviewSuperbuildBuild/install/lib/libavcodec.so.55.69.100", \
+  "/home/ubuntu/paraviewSuperbuildBuild/install/lib/libavutil.so.52.92.100", \
+  "/home/ubuntu/paraviewSuperbuildBuild/install/lib/libswscale.so.2.6.100", \
+  "/home/ubuntu/paraviewSuperbuildBuild/install/lib/python3.6", \
+  "/home/ubuntu/Paraview/"]
+
+
 FROM nsfcareer/femtech:production AS femtechprod
 
 FROM ubuntu:18.04
@@ -47,30 +66,46 @@ FROM ubuntu:18.04
 RUN apt-get update && \
   apt-get install -y  --no-install-recommends \
   openmpi-bin libopenblas-base openssh-client openssl \
-  libgl1 libxt6 xvfb jq ca-certificates curl zip unzip\
+  libgl1 libxt6 xvfb jq ca-certificates curl zip unzip \
+  libopengl0 libpython3.6 \
   && rm -rf /var/lib/apt/lists/* 
 
 RUN curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
 RUN unzip awscliv2.zip
 RUN ./aws/install
 
-
 # Setup home environment
 RUN useradd -rm -d /home/ubuntu -s /bin/bash -g root -G sudo -u 1000 ubuntu
 USER ubuntu
 WORKDIR /home/ubuntu
 
-RUN mkdir FemTechRun FemTechRun/results FemTechRun/results/vtu
+RUN mkdir FemTechRun FemTechRun/results FemTechRun/results/vtu lib
 
 COPY --from=femtechprod ["/home/ubuntu/FemTechRun/ex5", \
   "/home/ubuntu/FemTechRun/materials.dat", \
   "/home/ubuntu/FemTechRun/coarse_brain.inp", \
+  "/home/ubuntu/FemTechRun/simulationMovie.py", \
   "/home/ubuntu/FemTechRun/"]
 
 COPY --from=multiviewport ["/home/ubuntu/MultiViewPortRun/brain3.ply", \
   "/home/ubuntu/MultiViewPortRun/Br_color3.jpg", \
   "/home/ubuntu/MultiViewPortRun/MultipleViewPorts", \
   "/home/ubuntu/FemTechRun/"]
+
+COPY --from=multiviewport ["/home/ubuntu/Paraview/pvpython", \
+  "/home/ubuntu/FemTechRun/"]
+
+COPY --from=multiviewport ["/home/ubuntu/Paraview/site-packages", \
+  "/home/ubuntu/Paraview/libavformat.so.55.48.100", \
+  "/home/ubuntu/Paraview/libavcodec.so.55.69.100", \
+  "/home/ubuntu/Paraview/libavutil.so.52.92.100", \
+  "/home/ubuntu/Paraview/libswscale.so.2.6.100", \
+  "/home/ubuntu/lib/"]
+
+RUN cd lib;ln -s libswscale.so.2.6.100 libswscale.so.2;ln -s libavformat.so.55.48.100 libavformat.so.55;ln -s libavcodec.so.55.69.100 libavcodec.so.55;ln -s libavutil.so.52.92.100 libavutil.so.52
+
+ENV LD_LIBRARY_PATH /home/ubuntu/lib:${LD_LIBRARY_PATH}
+ENV PYTHONPATH /home/ubuntu/lib/_paraview.zip:/home/ubuntu/lib:/home/ubuntu/lib/_vtk.zip
 
 COPY --chown=ubuntu:root ./simulation.sh .
 RUN chmod +x simulation.sh
